refactor: fetch기능을 axios 로 바꾸고 별개의 파일로 분리
